feat: Add support for GO_LOG to set logging values (#29)
GO_LOG is a well known env used to set logging level. This PR allows the user to control which log level should be used without aving to change implementation of code. This can be useful in scenarios where one want to have more insight over the action that run